tests for new functionality for --output-directory and --checkpoint-name-format

+++ tests/cli/sst_output_dir_config-DEV.sh
@@ -1,12 +1,12 @@
 #!/bin/bash
 #EXT_TEST TEST_FILE_PARAM ALL
-#EXT_TEST TEST_FILE_DESC "Tests basic directory creation using --output-directory"
+#EXT_TEST TEST_FILE_DESC "Tests relocation of config files"
 #EXT_TEST DEP "coreTestElement.coreTestComponent"
 
 check_odir() {
     odir=$1
     rm -Rf $odir
-    sst --output-directory=$odir cli-sdl.py
+    sst --output-directory=$odir --output-config=sst_output_dir.config cli-sdl.py
     retVal=$?
     if [ $retVal -ne 0 ]; then
         echo "ERROR : $retVal"
@@ -16,11 +16,15 @@ check_odir() {
         echo "ERROR : Output directory ${odir} does not exist!"
         exit 1
     fi
+    if [ ! -e "$odir/sst_output_dir.config" ]; then
+        echo "ERROR missing checkpoint file $odir/sst_output_dir.config"
+        exit 2
+    fi
     rm -Rf $odir
     echo "${odir} OK"
 }
 
-tname=sst_output_dir
+tname=sst_output_dir_config
 
 # Check 1: 1 level subdirectory
 check_odir $tname.$$
